Splits ShapeDrawer into partial classes.
