--- Runtime/Scripts/LineDrawer.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0b31596a7ad092346a1e5a80b0b89c47
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
