--- Runtime/Scripts/MeshDrawer.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1c8de3e265e3c9f4d8686be519a9fd8a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
