+++ Runtime/Scripts/ShapeDrawer.Box.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d9d82aa7a0cb2b49b42edeb4564afc3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
