+++ Runtime/Scripts/ShapeDrawer.Capsule.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6e8d682687b912b479c4a2e1bb5eb037
+guid: 3e5302fed1e546c40ba2a85214ccfddc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
