+++ Runtime/Scripts/ShapeDrawer.Mesh.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4be89a561ba7f664c987611bfec6d7d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
