+++ Runtime/Scripts/ShapeDrawer.Rectangle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4acee286b390ef44addfb2315bf0d25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
