+++ Runtime/Scripts/ShapeDrawer.Text.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69e014170ee1efd4eb2f710d1e8ddad7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
