--- Runtime/Scripts/TextDrawer.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c6548912511e8fc46a53c24abc440ba1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
